adjust cart

+++ cosmetics-shop/src/main/java/com/cosmetics/myshop/controller/ApiController.java
@@ -10,8 +10,11 @@ import javax.print.attribute.standard.JobKOctets;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -75,6 +78,10 @@ public class ApiController {
 	@ResponseBody
 	@GetMapping("/cart")
 	public List<CartItemDTO> getCartItems(Authentication authentication) {
+		if(authentication == null) {
+			return null;			
+		}
+		
 		User user = (User) authentication.getPrincipal();
 		int userId = user.getUserId();
 
@@ -101,42 +108,53 @@ public class ApiController {
 
 	@ResponseBody
 	@PostMapping("/cart/add")
-	public void addToCart(@RequestBody CartItemDTO addToCartDTO, Authentication authentication) {
-		User user = (User) authentication.getPrincipal();
-		int userId = user.getUserId();
+	public ResponseEntity<Void> addToCart(@RequestBody CartItemDTO addToCartDTO, Authentication authentication) {
+	    if(authentication != null) {
+	        User user = (User) authentication.getPrincipal();
+	        int userId = user.getUserId();
 
-		ShoppingSession shoppingSession = shoppingSessionService.findShoppingSessionByUserId(userId);
-		if (shoppingSession == null) {
-			shoppingSession = new ShoppingSession();
-			shoppingSession.setUser(user);
-			shoppingSession.setCreatedAt(new Date());
-			shoppingSessionService.saveShoppingSession(shoppingSession);
-		}
+	        ShoppingSession shoppingSession = shoppingSessionService.findShoppingSessionByUserId(userId);
+	        if (shoppingSession == null) {
+	            shoppingSession = new ShoppingSession();
+	            shoppingSession.setUser(user);
+	            shoppingSession.setCreatedAt(new Date());
+	            shoppingSessionService.saveShoppingSession(shoppingSession);
+	        }
 
-		// Check if the product is already in the cart
-		CartItem existingCartItem = 
-				cartItemService.findCartItem(shoppingSession.getId(), addToCartDTO.getProductId());
-		if (existingCartItem != null) {
-			// Update quantity if the product is already in the cart
-			existingCartItem.setQuantity(existingCartItem.getQuantity() + addToCartDTO.getQuantity());
-			cartItemService.updateCartItemQuantity(shoppingSession.getId(), 
-								addToCartDTO.getProductId(), existingCartItem.getQuantity());
-			return;
-		}
+	        // Check if the product is already in the cart
+	        CartItem existingCartItem =
+	                cartItemService.findCartItem(shoppingSession.getId(), addToCartDTO.getProductId());
+	        if (existingCartItem != null) {
+	            // Update quantity if the product is already in the cart
+	            existingCartItem.setQuantity(existingCartItem.getQuantity() + addToCartDTO.getQuantity());
+	            cartItemService.updateCartItemQuantity(shoppingSession.getId(),
+	                    addToCartDTO.getProductId(), existingCartItem.getQuantity());
+	            return ResponseEntity.ok().build();
+	        }
 
-		// Create new cart item if the product is not in the cart
-		cartItemService.addToCart(shoppingSession.getId(), addToCartDTO.getProductId(), 
-									addToCartDTO.getQuantity(), new Date());
+	        cartItemService.addToCart(shoppingSession.getId(), addToCartDTO.getProductId(),
+	                addToCartDTO.getQuantity(), new Date());
+	        return ResponseEntity.ok().build();
+	    } 
+	    else {
+	        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+	    }
 	}
 	
 	@ResponseBody
-	@GetMapping("/cart/delete/{productId}")
-	public void deleteCartItem(@PathVariable Integer productId, Authentication authentication) {
-		User user = (User) authentication.getPrincipal();
-		int userId = user.getUserId();
+	@DeleteMapping("/cart/delete/{productId}")
+	public ResponseEntity<Void> deleteCartItem(@PathVariable Integer productId, Authentication authentication) {
+		if(authentication != null) {
+			User user = (User) authentication.getPrincipal();
+			int userId = user.getUserId();
 
-		ShoppingSession shoppingSession = shoppingSessionService.findShoppingSessionByUserId(userId);
+			ShoppingSession shoppingSession = shoppingSessionService.findShoppingSessionByUserId(userId);
 
-		cartItemService.deleteCartItem(shoppingSession.getId(), productId);
+			cartItemService.deleteCartItem(shoppingSession.getId(), productId);
+			 return ResponseEntity.ok().build();
+		}
+		else {
+	        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+	    }		
 	}
 }
